% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JointAItestpkg.R
\docType{package}
\name{JointAItestpkg}
\alias{JointAItestpkg-package}
\alias{JointAItestpkg}
\title{A testpackage for github actions}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Jane Doe \email{jane@example.com}

}
